Add CachedLLVMAliasIterator and implement FilteredLLVMAliasSet in terms of it
